requirements.txt
# API Client Dependencies
requests>=2.31.0
aiohttp>=3.9.0

# AWS SDK
boto3>=1.28.0

# Environment Configuration
python-dotenv>=1.0.0

# Image Processing (if needed for future features)
Pillow>=10.0.0

# JSON Schema Validation (optional but recommended for structured outputs)
jsonschema>=4.20.0

# CSV Processing
pandas>=2.0.0
